Move chip arg to be first in portMode()

+++ src/LapMCP23017.h
@@ -29,7 +29,7 @@ public:
   LapMCP23017(uint8_t, TwoWire *);
   void begin();
   void pinMode(uint8_t, uint8_t);
-  void portMode(uint16_t, uint16_t, uint8_t);
+  void portMode(uint8_t, uint16_t, uint16_t);
   uint8_t digitalRead(uint8_t);
   void digitalWrite(uint8_t, uint8_t);
   uint16_t read(uint8_t);
